last possible commit

+++ routes/ideas.js
@@ -47,18 +47,27 @@ router.post('/', async (req, res) => {
 //Upadte idea
 router.put('/:id', async (req, res) => {
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(
-            req.params.id,
-            {
-                $set: {
-                    text: req.body.text,
-                    tag: req.body.tag,
+        const idea = await Idea.findById(req.params.id);
+        if (idea.username === req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(
+                req.params.id,
+                {
+                    $set: {
+                        text: req.body.text,
+                        tag: req.body.tag,
+    
+                    }
+                },
+                { new: true }
+            );
+            return res.json({success: true, data: updatedIdea});
+        }
+        //Usernames does not match
 
-                }
-            },
-            { new: true }
-        );
-        res.json({success: true, data: updatedIdea});
+        console.log(error);
+        res.status(403).json({success: false, error: 'you are not authorized to update this resource' });
+
+
     } catch (error) {
         console.log(error);
         res.status(500).json({success: false, error: 'something went wrong' });
@@ -69,8 +78,17 @@ router.put('/:id', async (req, res) => {
 
     router.delete('/:id', async(req, res) => {
         try {
-            await Idea.findByIdAndDelete(req.params.id);
-            res.json({success: true, data: {} });
+            const idea = await Idea.findById(req.params.id);
+
+            //Match the usernames
+            if (idea.username === req.body.username) {
+                await Idea.findByIdAndDelete(req.params.id);
+                 return res.json({success: true, data: {} });
+            }
+
+            //username do not match 
+            console.log(error);
+        res.status(403).json({success: false, error: 'you are not authorized to delete this resource' });
         } catch (error) {
             console.log(error);
         res.status(500).json({success: false, error: 'something went wrong' });
